+++ src/main/resources/locks_by_type.sql
@@ -1,0 +1,4 @@
+select relation::regclass as relname, count(*) as count, mode
+from pg_locks
+where relation::regclass::text like :relname
+group by mode, relation